requirements.txt
python-jose[cryptography]
passlib[bcrypt]
fastapi[standard]
uvicorn[standard]
motor==3.6.0
redis==4.5.5
charset-normalizer==3.3.2
setuptools
python-multipart
authlib
itsdangerous
asyncio
httpx
websockets
slowapi
pandas
fastapi-sso
asgiref
pymongo
ruff
jinja2
pyjwt
markdown
pytesseract
pdf2image
opencv-python
psutil
pypdf
openai
langchain-community
langchain
langchain-openai
PyPDF2
scikit-learn
sentence-transformers
tiktoken
mkdocs
mkdocs-material

# Additional dependencies for Cerebras integration
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0